species_classifier/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
faiss-cpu>=1.7.0
Pillow>=8.3.0
pandas>=1.3.0
scikit-learn>=0.24.0
python-multipart>=0.0.5
tqdm>=4.62.0
matplotlib>=3.4.0
numpy>=1.21.0